Final review: extract hooks, card components, display extractors, dedup server guards
- Decompose creator/Onboarding.jsx (677→120 lines) into step components + useCreatorOnboardingForm hook
- Extract useDraftSubmission hook, reduce DraftUploadSection from 11 to 2 props
- Extract useBriefDisplayData hook from BriefDetail.jsx (371→330 lines)
- Extract CreatorBriefCard + CreatorProjectCard from Dashboard.jsx (387→159 lines)
- Add display name/photo extractors (utils/extractors.js) to eliminate Law of Demeter chains
- Extract findDraftForAction server helper to dedup approve/revision route guards
- Fix index keys across DraftStatusView, DraftHistory, ProjectStatusSection, Dashboard, BriefDetail
- Add WCAG a11y: Chip focus ring + 44px touch target
