+++ client/src/pages/creator/Dashboard.jsx
@@ -2,11 +2,12 @@ import { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useAuth } from '../../contexts/AuthContext';
 import { getCreatorStats, getProjects, getBriefs } from '../../api';
-import { formatCents, PROJECT_STATUS_LABELS } from '../../utils/constants';
+import { formatCents } from '../../utils/constants';
 import StatCard from '../../components/common/StatCard';
-import StatusBadge from '../../components/common/StatusBadge';
 import EmptyState from '../../components/common/EmptyState';
 import LoadingSpinner from '../../components/common/LoadingSpinner';
+import CreatorBriefCard from '../../components/creator/CreatorBriefCard';
+import CreatorProjectCard from '../../components/creator/CreatorProjectCard';
 
 export default function Dashboard() {
   const navigate = useNavigate();
@@ -56,7 +57,6 @@ export default function Dashboard() {
 
   return (
     <div className="max-w-5xl mx-auto px-4 py-8">
-      {/* Error banner */}
       {error && (
         <div className="bg-red-50 border border-red-200 rounded-xl p-4 mb-6">
           <p className="text-sm text-red-700 font-body">{error}</p>
@@ -70,23 +70,9 @@ export default function Dashboard() {
         </h1>
         {subtitle && (
           <p className="font-body text-muted text-sm flex items-center gap-1.5">
-            <svg
-              className="w-4 h-4 text-creatorAccent"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-              strokeWidth={1.5}
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                d="M15 10.5a3 3 0 11-6 0 3 3 0 016 0z"
-              />
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                d="M19.5 10.5c0 7.142-7.5 11.25-7.5 11.25S4.5 17.642 4.5 10.5a7.5 7.5 0 1115 0z"
-              />
+            <svg className="w-4 h-4 text-creatorAccent" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={1.5}>
+              <path strokeLinecap="round" strokeLinejoin="round" d="M15 10.5a3 3 0 11-6 0 3 3 0 016 0z" />
+              <path strokeLinecap="round" strokeLinejoin="round" d="M19.5 10.5c0 7.142-7.5 11.25-7.5 11.25S4.5 17.642 4.5 10.5a7.5 7.5 0 1115 0z" />
             </svg>
             {subtitle}
           </p>
@@ -99,61 +85,19 @@ export default function Dashboard() {
           label="Monthly Earnings"
           value={formatCents(monthlyEarnings)}
           creator
-          icon={
-            <svg
-              className="w-5 h-5"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-              strokeWidth={1.5}
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                d="M12 6v12m-3-2.818l.879.659c1.171.879 3.07.879 4.242 0 1.172-.879 1.172-2.303 0-3.182C13.536 12.219 12.768 12 12 12c-.725 0-1.45-.22-2.003-.659-1.106-.879-1.106-2.303 0-3.182s2.9-.879 4.006 0l.415.33M21 12a9 9 0 11-18 0 9 9 0 0118 0z"
-              />
-            </svg>
-          }
+          icon={<svg className="w-5 h-5" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={1.5}><path strokeLinecap="round" strokeLinejoin="round" d="M12 6v12m-3-2.818l.879.659c1.171.879 3.07.879 4.242 0 1.172-.879 1.172-2.303 0-3.182C13.536 12.219 12.768 12 12 12c-.725 0-1.45-.22-2.003-.659-1.106-.879-1.106-2.303 0-3.182s2.9-.879 4.006 0l.415.33M21 12a9 9 0 11-18 0 9 9 0 0118 0z" /></svg>}
         />
         <StatCard
           label="Active Projects"
           value={activeCount}
           creator
-          icon={
-            <svg
-              className="w-5 h-5"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-              strokeWidth={1.5}
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                d="M20.25 14.15v4.25c0 1.094-.787 2.036-1.872 2.18-2.087.277-4.216.42-6.378.42s-4.291-.143-6.378-.42c-1.085-.144-1.872-1.086-1.872-2.18v-4.25m16.5 0a2.18 2.18 0 00.75-1.661V8.706c0-1.081-.768-2.015-1.837-2.175a48.114 48.114 0 00-3.413-.387m4.5 8.006c-.194.165-.42.295-.673.38A23.978 23.978 0 0112 15.75c-2.648 0-5.195-.429-7.577-1.22a2.016 2.016 0 01-.673-.38m0 0A2.18 2.18 0 013 12.489V8.706c0-1.081.768-2.015 1.837-2.175a48.111 48.111 0 013.413-.387m7.5 0V5.25A2.25 2.25 0 0013.5 3h-3a2.25 2.25 0 00-2.25 2.25v.894m7.5 0a48.667 48.667 0 00-7.5 0"
-              />
-            </svg>
-          }
+          icon={<svg className="w-5 h-5" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={1.5}><path strokeLinecap="round" strokeLinejoin="round" d="M20.25 14.15v4.25c0 1.094-.787 2.036-1.872 2.18-2.087.277-4.216.42-6.378.42s-4.291-.143-6.378-.42c-1.085-.144-1.872-1.086-1.872-2.18v-4.25m16.5 0a2.18 2.18 0 00.75-1.661V8.706c0-1.081-.768-2.015-1.837-2.175a48.114 48.114 0 00-3.413-.387m4.5 8.006c-.194.165-.42.295-.673.38A23.978 23.978 0 0112 15.75c-2.648 0-5.195-.429-7.577-1.22a2.016 2.016 0 01-.673-.38m0 0A2.18 2.18 0 013 12.489V8.706c0-1.081.768-2.015 1.837-2.175a48.111 48.111 0 013.413-.387m7.5 0V5.25A2.25 2.25 0 0013.5 3h-3a2.25 2.25 0 00-2.25 2.25v.894m7.5 0a48.667 48.667 0 00-7.5 0" /></svg>}
         />
         <StatCard
           label="New Briefs"
           value={newBriefsCount}
           creator
-          icon={
-            <svg
-              className="w-5 h-5"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-              strokeWidth={1.5}
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                d="M21.75 6.75v10.5a2.25 2.25 0 01-2.25 2.25h-15a2.25 2.25 0 01-2.25-2.25V6.75m19.5 0A2.25 2.25 0 0019.5 4.5h-15a2.25 2.25 0 00-2.25 2.25m19.5 0v.243a2.25 2.25 0 01-1.07 1.916l-7.5 4.615a2.25 2.25 0 01-2.36 0L3.32 8.91a2.25 2.25 0 01-1.07-1.916V6.75"
-              />
-            </svg>
-          }
+          icon={<svg className="w-5 h-5" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={1.5}><path strokeLinecap="round" strokeLinejoin="round" d="M21.75 6.75v10.5a2.25 2.25 0 01-2.25 2.25h-15a2.25 2.25 0 01-2.25-2.25V6.75m19.5 0A2.25 2.25 0 0019.5 4.5h-15a2.25 2.25 0 00-2.25 2.25m19.5 0v.243a2.25 2.25 0 01-1.07 1.916l-7.5 4.615a2.25 2.25 0 01-2.36 0L3.32 8.91a2.25 2.25 0 01-1.07-1.916V6.75" /></svg>}
         />
       </div>
 
@@ -166,121 +110,20 @@ export default function Dashboard() {
         {briefs.length === 0 ? (
           <div className="card">
             <EmptyState
-              icon={
-                <svg
-                  className="w-7 h-7"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                  strokeWidth={1.5}
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="M21.75 6.75v10.5a2.25 2.25 0 01-2.25 2.25h-15a2.25 2.25 0 01-2.25-2.25V6.75m19.5 0A2.25 2.25 0 0019.5 4.5h-15a2.25 2.25 0 00-2.25 2.25m19.5 0v.243a2.25 2.25 0 01-1.07 1.916l-7.5 4.615a2.25 2.25 0 01-2.36 0L3.32 8.91a2.25 2.25 0 01-1.07-1.916V6.75"
-                  />
-                </svg>
-              }
+              icon={<svg className="w-7 h-7" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={1.5}><path strokeLinecap="round" strokeLinejoin="round" d="M21.75 6.75v10.5a2.25 2.25 0 01-2.25 2.25h-15a2.25 2.25 0 01-2.25-2.25V6.75m19.5 0A2.25 2.25 0 0019.5 4.5h-15a2.25 2.25 0 00-2.25 2.25m19.5 0v.243a2.25 2.25 0 01-1.07 1.916l-7.5 4.615a2.25 2.25 0 01-2.36 0L3.32 8.91a2.25 2.25 0 01-1.07-1.916V6.75" /></svg>}
               title="No new briefs yet"
               description="When brands match with your style, briefs will appear here."
             />
           </div>
         ) : (
           <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-            {briefs.map((brief) => {
-              const contentType =
-                brief.contentRequest?.contentType ||
-                brief.contentType ||
-                brief.request?.contentType ||
-                'Content Project';
-              const neighborhood =
-                brief.brand?.neighborhood ||
-                brief.neighborhood ||
-                brief.request?.neighborhood ||
-                '';
-              const pay =
-                brief.price ?? brief.pay ?? brief.request?.budget ?? brief.budget ?? 0;
-              const deliverables =
-                brief.deliverables ||
-                brief.request?.deliverables ||
-                [];
-              const timeline =
-                brief.timeline || brief.request?.timeline || '';
-              const matchId = brief.matchId || brief.id || brief._id;
-
-              return (
-                <button
-                  key={matchId}
-                  onClick={() => navigate(`/creator/brief/${matchId}`)}
-                  className="card text-left hover:shadow-md hover:border-creatorAccent/30 transition-all duration-200 group"
-                >
-                  <div className="flex items-start justify-between mb-3">
-                    <div className="flex-1 min-w-0">
-                      <p className="font-display text-lg font-semibold text-dark truncate">
-                        {neighborhood ? `${neighborhood} \u00B7 ` : ''}
-                        {contentType}
-                      </p>
-                      {timeline && (
-                        <p className="font-body text-xs text-muted mt-0.5">
-                          {timeline}
-                        </p>
-                      )}
-                    </div>
-                    <span className="text-xs text-muted font-body">Why this brief?</span>
-                  </div>
-
-                  {/* Deliverables */}
-                  {deliverables && (
-                    <div className="flex flex-wrap gap-1.5 mb-3">
-                      {Array.isArray(deliverables) ? (
-                        <>
-                          {deliverables.slice(0, 3).map((d, i) => (
-                            <span
-                              key={i}
-                              className="px-2 py-0.5 text-xs rounded-full bg-creatorLight text-creator font-medium"
-                            >
-                              {d}
-                            </span>
-                          ))}
-                          {deliverables.length > 3 && (
-                            <span className="px-2 py-0.5 text-xs rounded-full bg-bgWarm text-muted font-medium">
-                              +{deliverables.length - 3} more
-                            </span>
-                          )}
-                        </>
-                      ) : (
-                        <span className="px-2 py-0.5 text-xs rounded-full bg-creatorLight text-creator font-medium">
-                          {deliverables}
-                        </span>
-                      )}
-                    </div>
-                  )}
-
-                  {/* Pay — prominent */}
-                  <div className="flex items-center justify-between mt-auto pt-3 border-t border-border">
-                    <span className="font-display text-2xl font-bold text-dark">
-                      {formatCents(pay)}
-                    </span>
-                    <span className="font-body text-sm text-creatorAccent font-semibold flex items-center gap-1 group-hover:gap-2 transition-all duration-200">
-                      View Brief
-                      <svg
-                        className="w-4 h-4 group-hover:translate-x-0.5 transition-transform duration-200"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                        stroke="currentColor"
-                        strokeWidth={2}
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          d="M13.5 4.5L21 12m0 0l-7.5 7.5M21 12H3"
-                        />
-                      </svg>
-                    </span>
-                  </div>
-                </button>
-              );
-            })}
+            {briefs.map((brief) => (
+              <CreatorBriefCard
+                key={brief.matchId || brief.id || brief._id}
+                brief={brief}
+                onClick={() => navigate(`/creator/brief/${brief.matchId || brief.id || brief._id}`)}
+              />
+            ))}
           </div>
         )}
       </section>
@@ -294,91 +137,20 @@ export default function Dashboard() {
         {projects.length === 0 ? (
           <div className="card">
             <EmptyState
-              icon={
-                <svg
-                  className="w-7 h-7"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                  strokeWidth={1.5}
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="M2.25 15.75l5.159-5.159a2.25 2.25 0 013.182 0l5.159 5.159m-1.5-1.5l1.409-1.409a2.25 2.25 0 013.182 0l2.909 2.909M3 21h18M3 21a2.25 2.25 0 01-2.25-2.25V5.25A2.25 2.25 0 013 3h18a2.25 2.25 0 012.25 2.25v13.5A2.25 2.25 0 0121 21"
-                  />
-                </svg>
-              }
+              icon={<svg className="w-7 h-7" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={1.5}><path strokeLinecap="round" strokeLinejoin="round" d="M2.25 15.75l5.159-5.159a2.25 2.25 0 013.182 0l5.159 5.159m-1.5-1.5l1.409-1.409a2.25 2.25 0 013.182 0l2.909 2.909M3 21h18M3 21a2.25 2.25 0 01-2.25-2.25V5.25A2.25 2.25 0 013 3h18a2.25 2.25 0 012.25 2.25v13.5A2.25 2.25 0 0121 21" /></svg>}
               title="No active projects"
               description="When you accept a brief, your projects will appear here."
             />
           </div>
         ) : (
           <div className="space-y-3">
-            {projects.map((project) => {
-              const projectId = project.id || project._id;
-              const brandName =
-                project.brandProfile?.user?.name ||
-                project.brandProfile?.businessName ||
-                project.brand?.name ||
-                project.brandName ||
-                'Brand';
-              const contentType =
-                project.match?.contentRequest?.contentType ||
-                project.contentType ||
-                project.request?.contentType ||
-                'Content Project';
-              const status = project.status || 'BRIEF_SENT';
-              const pay = project.price ?? project.pay ?? project.budget ?? 0;
-
-              return (
-                <button
-                  key={projectId}
-                  onClick={() => navigate(`/creator/project/${projectId}`)}
-                  className="card w-full text-left hover:shadow-md hover:border-creatorAccent/30 transition-all duration-200 group"
-                >
-                  <div className="flex items-center gap-4">
-                    {/* Brand initial */}
-                    <div className="w-11 h-11 rounded-xl bg-creatorLight flex items-center justify-center shrink-0">
-                      <span className="font-display text-lg font-bold text-creator">
-                        {brandName.charAt(0)}
-                      </span>
-                    </div>
-
-                    {/* Info */}
-                    <div className="flex-1 min-w-0">
-                      <p className="font-body font-semibold text-dark truncate">
-                        {brandName}
-                      </p>
-                      <p className="font-body text-sm text-muted truncate">
-                        {contentType}
-                      </p>
-                    </div>
-
-                    {/* Status + Pay */}
-                    <div className="flex items-center gap-4 shrink-0">
-                      <StatusBadge status={status} />
-                      <span className="font-display text-lg font-bold text-dark">
-                        {formatCents(pay)}
-                      </span>
-                      <svg
-                        className="w-5 h-5 text-muted group-hover:text-creatorAccent group-hover:translate-x-0.5 transition-all duration-200"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                        stroke="currentColor"
-                        strokeWidth={2}
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          d="M8.25 4.5l7.5 7.5-7.5 7.5"
-                        />
-                      </svg>
-                    </div>
-                  </div>
-                </button>
-              );
-            })}
+            {projects.map((project) => (
+              <CreatorProjectCard
+                key={project.id || project._id}
+                project={project}
+                onClick={() => navigate(`/creator/project/${project.id || project._id}`)}
+              />
+            ))}
           </div>
         )}
       </section>
